Added precision in %i, d, u, p. Added some reminders on my phone

+++ libft/ft_putnbr_fd.c
@@ -6,7 +6,7 @@
 /*   By: xvoorvaa <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/11/20 14:58:40 by xvoorvaa      #+#    #+#                 */
-/*   Updated: 2021/06/07 14:38:04 by xvoorvaa      ########   odam.nl         */
+/*   Updated: 2021/06/12 16:20:27 by xvoorvaa      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_putnbr_fd(int n, int fd, size_t len)
 {
-	long	i;
+	int	i;
 
 	i = n;
 	if (i < 0)
